third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/search-styles-iso-8859-8-expected.txt

This is a testharness.js-based test.
[FAIL] <search> - unicode-bidi
  assert_equals: expected "bidi-override" but got "normal"
Harness: the test ran to completion.
